pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "icm-system"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "google-cloud-bigquery>=3.0",
    "flask>=3.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "langgraph>=0.2",
    "requests>=2.31",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "tools*", "db*"]
